restructure qredo
