+++ apps/wallet/src/background/index.ts
@@ -8,9 +8,9 @@ import { LOCK_ALARM_NAME } from './Alarms';
 import FeatureGating from './FeatureGating';
 import NetworkEnv from './NetworkEnv';
 import Permissions from './Permissions';
-import * as Qredo from './Qredo';
 import { Connections } from './connections';
 import Keyring from './keyring';
+import * as Qredo from './qredo';
 import { isSessionStorageSupported } from './storage-utils';
 import { openInNewTab } from '_shared/utils';
 import { MSG_CONNECT } from '_src/content-script/keep-bg-alive';
